✨ Count nodes per area
